# apps/daemon/pkg/toolbox/docs/swagger.yaml
definitions:
...
  Chart:
    properties:
      elements:
        items:
          $ref: '#/definitions/ChartElement'
        type: array
      png:
        type: string
      title:
        type: string
      type:
        type: string
      x_label:
        type: string
      x_scale:
        type: string
      x_tick_labels:
        items:
          type: string
        type: array
      x_ticks:
        items:
          type: number
        type: array
      y_label:
        type: string
      y_scale:
        type: string
      y_tick_labels:
        items:
          type: string
        type: array
      y_ticks:
        items:
          type: number
        type: array
    type: object
  ChartElement:
    properties:
      angle:
        type: number
      first_quartile:
        type: number
      group:
        type: string
      label:
        type: string
      max:
        type: number
      median:
        type: number
      min:
        type: number
      outliers:
        items:
          type: number
        type: array
      png:
        type: string
      points:
        items:
          items:
            type: number
          type: array
        type: array
      radius:
        type: number
      third_quartile:
        type: number
      title:
        type: string
      type:
        type: string
      value:
        type: string
      x_label:
        type: string
      y_label:
        type: string
    type: object
  CodeRunArtifacts:
    properties:
      charts:
        items:
          $ref: '#/definitions/Chart'
        type: array
    type: object
  CodeRunRequest:
    properties:
      argv:
        items:
          type: string
        type: array
      code:
        type: string
      envs:
        additionalProperties:
          type: string
        type: object
      language:
        description: python, javascript, typescript
        type: string
      timeout:
        type: integer
    required:
      - code
      - language
    type: object
  CodeRunResponse:
    properties:
      artifacts:
        $ref: '#/definitions/CodeRunArtifacts'
      exitCode:
        type: integer
      result:
        type: string
    type: object
...
  DaemonErrorCode:
    enum:
      - GIT_AUTH_FAILED
      - GIT_AUTH_FORBIDDEN
      - GIT_REPO_NOT_FOUND
      - GIT_BRANCH_NOT_FOUND
      - GIT_REF_NOT_FOUND
      - GIT_EMPTY_REPO
      - GIT_PUSH_REJECTED
      - GIT_DIRTY_WORKTREE
      - GIT_BRANCH_EXISTS
      - GIT_MERGE_CONFLICT
      - GIT_REPO_EXISTS
      - FILE_NOT_FOUND
      - FILE_ACCESS_DENIED
      - INVALID_FILE_PATH
      - LSP_SERVER_NOT_INITIALIZED
      - LSP_INVALID_REQUEST
      - PROCESS_EXECUTION_TIMEOUT
      - PROCESS_INVALID_COMMAND
      - PROCESS_NOT_FOUND
      - SESSION_ENDED
    type: string
    x-enum-varnames:
      - CodeGitAuthFailed
      - CodeGitAuthForbidden
      - CodeGitRepoNotFound
      - CodeGitBranchNotFound
      - CodeGitRefNotFound
      - CodeGitEmptyRepo
      - CodeGitPushRejected
      - CodeGitDirtyWorktree
      - CodeGitBranchExists
      - CodeGitMergeConflict
      - CodeGitRepoExists
      - CodeFileNotFound
      - CodeFileAccessDenied
      - CodeInvalidFilePath
      - CodeLspServerNotInitialized
      - CodeLspInvalidRequest
      - CodeProcessExecutionTimeout
      - CodeProcessInvalidCommand
      - CodeProcessNotFound
      - CodeSessionEnded
...
  ErrorResponse:
    description: Error response
    properties:
      code:
        allOf:
          - $ref: '#/definitions/DaemonErrorCode'
        example: GIT_REPO_NOT_FOUND
      message:
        example: Bad request
        type: string
      method:
        example: GET
        type: string
      path:
        example: /api/resource
        type: string
      source:
        example: DAYTONA_DAEMON
        type: string
      statusCode:
        example: 400
        type: integer
      timestamp:
        example: '2023-01-01T12:00:00Z'
        type: string
    required:
      - message
      - path
      - statusCode
      - timestamp
    type: object
...
  FileStatus:
    properties:
      extra:
        type: string
      name:
        type: string
      staging:
        $ref: '#/definitions/Status'
      worktree:
        $ref: '#/definitions/Status'
    required:
      - extra
      - name
      - staging
      - worktree
    type: object
...
  GitStatus:
    properties:
      ahead:
        type: integer
      behind:
        type: integer
      branchPublished:
        type: boolean
      currentBranch:
        type: string
      fileStatus:
        items:
          $ref: '#/definitions/FileStatus'
        type: array
    required:
      - currentBranch
      - fileStatus
    type: object
  InitializeRequest:
    properties:
      token:
        type: string
    required:
      - token
    type: object
...
  ListRecordingsResponse:
    properties:
      recordings:
        items:
          $ref: '#/definitions/Recording'
        type: array
    required:
      - recordings
    type: object
...
  ProcessErrorsResponse:
    properties:
      errors:
        type: string
      processName:
        type: string
    type: object
  ProcessLogsResponse:
    properties:
      logs:
        type: string
      processName:
        type: string
    type: object
  ProcessRestartResponse:
    properties:
      message:
        type: string
      processName:
        type: string
    type: object
...
  ProcessStatusResponse:
    properties:
      processName:
        type: string
      running:
        type: boolean
    type: object
...
  Recording:
    properties:
      durationSeconds:
        type: number
      endTime:
        type: string
      fileName:
        type: string
      filePath:
        type: string
      id:
        type: string
      sizeBytes:
        type: integer
      startTime:
        type: string
      status:
        type: string
    required:
      - fileName
      - filePath
      - id
      - startTime
      - status
    type: object
...
  Status:
    enum:
      - Unmodified
      - Untracked
      - Modified
      - Added
      - Deleted
      - Renamed
      - Copied
      - Updated but unmerged
    type: string
    x-enum-varnames:
      - Unmodified
      - Untracked
      - Modified
      - Added
      - Deleted
      - Renamed
      - Copied
      - UpdatedButUnmerged
...
info:
  contact: {}
  description: Daytona Toolbox API
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Toolbox API
  version: v0.0.0-dev
paths:
...
  /computeruse/process/{processName}/errors:
    get:
      description: Get errors for a specific computer use process
      operationId: GetProcessErrors
      parameters:
        - description: Process name to get errors for
          in: path
          name: processName
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProcessErrorsResponse'
      summary: Get process errors
      tags:
        - computer-use
  /computeruse/process/{processName}/logs:
    get:
      description: Get logs for a specific computer use process
      operationId: GetProcessLogs
      parameters:
        - description: Process name to get logs for
          in: path
          name: processName
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProcessLogsResponse'
      summary: Get process logs
      tags:
        - computer-use
  /computeruse/process/{processName}/restart:
    post:
      description: Restart a specific computer use process
      operationId: RestartProcess
      parameters:
        - description: Process name to restart
          in: path
          name: processName
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProcessRestartResponse'
      summary: Restart specific process
      tags:
        - computer-use
  /computeruse/process/{processName}/status:
    get:
      description: Check if a specific computer use process is running
      operationId: GetProcessStatus
      parameters:
        - description: Process name to check
          in: path
          name: processName
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProcessStatusResponse'
      summary: Get specific process status
      tags:
        - computer-use
  /computeruse/recordings:
    get:
      description: Get a list of all recordings (active and completed)
      operationId: ListRecordings
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ListRecordingsResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List all recordings
      tags:
        - computer-use
  /computeruse/recordings/{id}:
    delete:
      description: Delete a recording file by ID
      operationId: DeleteRecording
      parameters:
        - description: Recording ID
          in: path
          name: id
          required: true
          type: string
      responses:
        '204':
          description: No Content
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '409':
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Delete a recording
      tags:
        - computer-use
    get:
      description: Get details of a specific recording by ID
      operationId: GetRecording
      parameters:
        - description: Recording ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Recording'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get recording details
      tags:
        - computer-use
...
  /git/status:
    get:
      description: Get the Git status of the repository at the specified path
      operationId: GetStatus
      parameters:
        - description: Repository path
          in: query
          name: path
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/GitStatus'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '404':
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get Git status
      tags:
        - git
  /init:
    post:
      description: Set the auth token and initialize telemetry for the toolbox server
      operationId: Initialize
      parameters:
        - description: Initialization request
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/InitializeRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initialize toolbox server
      tags:
        - server
...
  /process/code-run:
    post:
      consumes:
        - application/json
      description: Execute Python, JavaScript, or TypeScript code and return output,
        exit code, and artifacts
      operationId: CodeRun
      parameters:
        - description: Code execution request
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/CodeRunRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/CodeRunResponse'
        '400':
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        '408':
          description: Request Timeout
          schema:
            $ref: '#/definitions/ErrorResponse'
        '500':
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Execute code
      tags:
        - process
...
swagger: '2.0'