# libs/toolbox-api-client-go/model_daemon_error_code.go
/*
Daytona Toolbox API

Daytona Toolbox API

API version: v0.0.0-dev
*/

// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

package toolbox

import (
	"encoding/json"
)

// DaemonErrorCode the model 'DaemonErrorCode'
type DaemonErrorCode string

// List of DaemonErrorCode
const (
	DAEMONERRORCODE_CodeGitAuthFailed DaemonErrorCode = "GIT_AUTH_FAILED"
	DAEMONERRORCODE_CodeGitAuthForbidden DaemonErrorCode = "GIT_AUTH_FORBIDDEN"
	DAEMONERRORCODE_CodeGitRepoNotFound DaemonErrorCode = "GIT_REPO_NOT_FOUND"
	DAEMONERRORCODE_CodeGitBranchNotFound DaemonErrorCode = "GIT_BRANCH_NOT_FOUND"
	DAEMONERRORCODE_CodeGitRefNotFound DaemonErrorCode = "GIT_REF_NOT_FOUND"
	DAEMONERRORCODE_CodeGitEmptyRepo DaemonErrorCode = "GIT_EMPTY_REPO"
	DAEMONERRORCODE_CodeGitPushRejected DaemonErrorCode = "GIT_PUSH_REJECTED"
	DAEMONERRORCODE_CodeGitDirtyWorktree DaemonErrorCode = "GIT_DIRTY_WORKTREE"
	DAEMONERRORCODE_CodeGitBranchExists DaemonErrorCode = "GIT_BRANCH_EXISTS"
	DAEMONERRORCODE_CodeGitMergeConflict DaemonErrorCode = "GIT_MERGE_CONFLICT"
	DAEMONERRORCODE_CodeGitRepoExists DaemonErrorCode = "GIT_REPO_EXISTS"
	DAEMONERRORCODE_CodeFileNotFound DaemonErrorCode = "FILE_NOT_FOUND"
	DAEMONERRORCODE_CodeFileAccessDenied DaemonErrorCode = "FILE_ACCESS_DENIED"
	DAEMONERRORCODE_CodeInvalidFilePath DaemonErrorCode = "INVALID_FILE_PATH"
	DAEMONERRORCODE_CodeLspServerNotInitialized DaemonErrorCode = "LSP_SERVER_NOT_INITIALIZED"
	DAEMONERRORCODE_CodeLspInvalidRequest DaemonErrorCode = "LSP_INVALID_REQUEST"
	DAEMONERRORCODE_CodeProcessExecutionTimeout DaemonErrorCode = "PROCESS_EXECUTION_TIMEOUT"
	DAEMONERRORCODE_CodeProcessInvalidCommand DaemonErrorCode = "PROCESS_INVALID_COMMAND"
	DAEMONERRORCODE_CodeProcessNotFound DaemonErrorCode = "PROCESS_NOT_FOUND"
	DAEMONERRORCODE_CodeSessionEnded DaemonErrorCode = "SESSION_ENDED"
	DAEMONERRORCODE_UNKNOWN_DEFAULT_OPEN_API DaemonErrorCode = "11184809"
)

// All allowed values of DaemonErrorCode enum
var AllowedDaemonErrorCodeEnumValues = []DaemonErrorCode{
	"GIT_AUTH_FAILED",
	"GIT_AUTH_FORBIDDEN",
	"GIT_REPO_NOT_FOUND",
	"GIT_BRANCH_NOT_FOUND",
	"GIT_REF_NOT_FOUND",
	"GIT_EMPTY_REPO",
	"GIT_PUSH_REJECTED",
	"GIT_DIRTY_WORKTREE",
	"GIT_BRANCH_EXISTS",
	"GIT_MERGE_CONFLICT",
	"GIT_REPO_EXISTS",
	"FILE_NOT_FOUND",
	"FILE_ACCESS_DENIED",
	"INVALID_FILE_PATH",
	"LSP_SERVER_NOT_INITIALIZED",
	"LSP_INVALID_REQUEST",
	"PROCESS_EXECUTION_TIMEOUT",
	"PROCESS_INVALID_COMMAND",
	"PROCESS_NOT_FOUND",
	"SESSION_ENDED",
	"11184809",
}

func (v *DaemonErrorCode) UnmarshalJSON(src []byte) error {
	var value string
	err := json.Unmarshal(src, &value)
	if err != nil {
		return err
	}
	enumTypeValue := DaemonErrorCode(value)
	for _, existing := range AllowedDaemonErrorCodeEnumValues {
		if existing == enumTypeValue {
			*v = enumTypeValue
			return nil
		}
	}

	*v = DAEMONERRORCODE_UNKNOWN_DEFAULT_OPEN_API
	return nil
}

// NewDaemonErrorCodeFromValue returns a pointer to a valid DaemonErrorCode
// for the value passed as argument, or an error if the value passed is not allowed by the enum
func NewDaemonErrorCodeFromValue(v string) (*DaemonErrorCode, error) {
	ev := DaemonErrorCode(v)
	if ev.IsValid() {
		return &ev, nil
	} else {
		enumValue := DAEMONERRORCODE_UNKNOWN_DEFAULT_OPEN_API
		return &enumValue, nil
	}
}

// IsValid return true if the value is valid for the enum, false otherwise
func (v DaemonErrorCode) IsValid() bool {
	for _, existing := range AllowedDaemonErrorCodeEnumValues {
		if existing == v {
			return true
		}
	}
	return false
}

// Ptr returns reference to DaemonErrorCode value
func (v DaemonErrorCode) Ptr() *DaemonErrorCode {
	return &v
}

type NullableDaemonErrorCode struct {
	value *DaemonErrorCode
	isSet bool
}

func (v NullableDaemonErrorCode) Get() *DaemonErrorCode {
	return v.value
}

func (v *NullableDaemonErrorCode) Set(val *DaemonErrorCode) {
	v.value = val
	v.isSet = true
}

func (v NullableDaemonErrorCode) IsSet() bool {
	return v.isSet
}

func (v *NullableDaemonErrorCode) Unset() {
	v.value = nil
	v.isSet = false
}

func NewNullableDaemonErrorCode(val *DaemonErrorCode) *NullableDaemonErrorCode {
	return &NullableDaemonErrorCode{value: val, isSet: true}
}

func (v NullableDaemonErrorCode) MarshalJSON() ([]byte, error) {
	return json.Marshal(v.value)
}

func (v *NullableDaemonErrorCode) UnmarshalJSON(src []byte) error {
	v.isSet = true
	return json.Unmarshal(src, &v.value)
}