# libs/toolbox-api-client-python-async/daytona_toolbox_api_client_async/models/daemon_error_code.py
# coding: utf-8

"""
    Daytona Toolbox API

    Daytona Toolbox API

    The version of the OpenAPI document: v0.0.0-dev
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
from enum import Enum
from typing_extensions import Self


class DaemonErrorCode(str, Enum):
    """
    DaemonErrorCode
    """

    """
    allowed enum values
    """
    CodeGitAuthFailed = 'GIT_AUTH_FAILED'
    CodeGitAuthForbidden = 'GIT_AUTH_FORBIDDEN'
    CodeGitRepoNotFound = 'GIT_REPO_NOT_FOUND'
    CodeGitBranchNotFound = 'GIT_BRANCH_NOT_FOUND'
    CodeGitRefNotFound = 'GIT_REF_NOT_FOUND'
    CodeGitEmptyRepo = 'GIT_EMPTY_REPO'
    CodeGitPushRejected = 'GIT_PUSH_REJECTED'
    CodeGitDirtyWorktree = 'GIT_DIRTY_WORKTREE'
    CodeGitBranchExists = 'GIT_BRANCH_EXISTS'
    CodeGitMergeConflict = 'GIT_MERGE_CONFLICT'
    CodeGitRepoExists = 'GIT_REPO_EXISTS'
    CodeFileNotFound = 'FILE_NOT_FOUND'
    CodeFileAccessDenied = 'FILE_ACCESS_DENIED'
    CodeInvalidFilePath = 'INVALID_FILE_PATH'
    CodeLspServerNotInitialized = 'LSP_SERVER_NOT_INITIALIZED'
    CodeLspInvalidRequest = 'LSP_INVALID_REQUEST'
    CodeProcessExecutionTimeout = 'PROCESS_EXECUTION_TIMEOUT'
    CodeProcessInvalidCommand = 'PROCESS_INVALID_COMMAND'
    CodeProcessNotFound = 'PROCESS_NOT_FOUND'
    CodeSessionEnded = 'SESSION_ENDED'
    UNKNOWN_DEFAULT_OPEN_API = 'unknown_default_open_api'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of DaemonErrorCode from a JSON string"""
        return cls(json.loads(json_str))


    @classmethod
    def _missing_(cls, value):
        return cls.UNKNOWN_DEFAULT_OPEN_API